pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docmost-cli"
version = "0.1.0"
description = "CLI tool for interacting with the Docmost API"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Gateway" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
docmost = "docmost.cli:cli"

[project.urls]
Homepage = "https://gitea.roboalch.com/gateway/docmost-cli"
Repository = "https://gitea.roboalch.com/gateway/docmost-cli"

[tool.hatch.build.targets.wheel]
packages = ["src/docmost"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
